DC Eco-label Data
Both csv files come from Open Data DC. The Certificate of Occupancy file gives the type of every business in DC so we can decide on our set of ecolabels. The Building Energy Benchmarking file gives the energy and water use for all business in DC that are larger than 50,000 sq ft within the District of Columbia, so we can decide on our minimum standard for sustainability.